removed version 1 of georouting
removed version 1 of georouting test
