+++ features/milo/georouting.spec.js
@@ -3,9 +3,33 @@ module.exports = {
   features: [
     {
       name: '@georouting',
-      path: '/customer-success-stories/princess-cruises-case-study',
-      envs: '@bacom_live',
+      path: '/test/features/blocks/georouting',
+      envs: '@milo_live',
+      tags: '@georouting @georouting-close @georouting-multi',
+    },
+    {
+      name: '@georouting',
+      path: '/de/test/features/blocks/georouting',
+      envs: '@milo_live',
       tags: '@georouting',
+    },
+    {
+      name: '@georouting',
+      path: '/test/features/blocks/georouting-off',
+      envs: '@milo_live',
+      tags: '@georouting-off',
+    },
+    {
+      name: '@georouting',
+      path: '/test/features/blocks/georouting-fallback-off',
+      envs: '@milo_live',
+      tags: '@fallback-off',
+    },
+    {
+      name: '@georouting',
+      path: '/test/features/blocks/georouting-fallback',
+      envs: '@milo_live',
+      tags: '@fallback-on',
     },
   ],
 };
